renamed project to vcdeploy = version controlled deployment; big update for rollout command

+++ plugins/plugin_reset-dir.class.php
@@ -17,9 +17,9 @@
  * License for the specific language governing rights and limitations
  * under the License.
  *
- * @package  sldeploy
+ * @package  vcdeploy
  * @author  Alexander Meindl
- * @link    https://github.com/alexandermeindl/sldeploy
+ * @link    https://github.com/alexandermeindl/vcdeploy
  */
 
 $plugin['info'] = 'Reset directory. If no project is specified, all active projects files/directories will be reseted.';
@@ -32,14 +32,28 @@ $plugin['options']['project'] = array(
                               'description' => 'Only reset data of this project',
                             );
 
-class SldeployPluginResetDir extends Sldeploy {
+$plugin['options']['with_backup'] = array(
+                                          'short_name'  => '-b',
+                                          'long_name'   => '--with_backup',
+                                          'action'      => 'StoreTrue',
+                                          'description' => 'Create a backup before the sync (default with setting without_backup=FALSE)',
+);
+
+$plugin['options']['without_backup'] = array(
+                                          'short_name'  => '-B',
+                                          'long_name'   => '--without_backup',
+                                          'action'      => 'StoreTrue',
+                                          'description' => 'Do not create a backup before the sync (default with setting without_backup=TRUE)',
+                                        );
+
+class VcdeployPluginResetDir extends Vcdeploy implements IVcdeployPlugin {
 
   /**
    * This function is run with the command
    *
    * @return int
    * @throws Exception
-   * @see sldeploy#run()
+   * @see vcdeploy#run()
    */
   public function run() {
 
@@ -47,7 +61,9 @@ class SldeployPluginResetDir extends Sldeploy {
     $this->validate_projects();
 
     // check backup directory if exists and is writable
-    $this->prepare_backup_dir();
+    if ($this->is_backup_required()) {
+      $this->prepare_backup_dir();
+    }
 
     if (isset($this->paras->command->options['project']) && !empty($this->paras->command->options['project'])) {
       $project_name = $this->paras->command->options['project'];
@@ -76,7 +92,7 @@ class SldeployPluginResetDir extends Sldeploy {
    * @param int $init initial value of counter
    *
    * @return int amount of working steps of this plugin
-   * @see Sldeploy#progressbar_init()
+   * @see Vcdeploy#progressbar_init()
    */
   public function get_steps($init = 0) {
     return $init++;
@@ -98,10 +114,17 @@ class SldeployPluginResetDir extends Sldeploy {
         if (!empty($tar_file)) {
 
           // create backup of existing data
-          $this->create_project_data_backup();
+          if ($this->is_backup_required()) {
+            $this->create_project_data_backup();
+          }
+          else {
+            $this->msg('Backup deactivated.');
+          }
 
           // remove existing target directory
-          $this->remove_directory($dir);
+          if (file_exists($dir)) {
+            $this->remove_directory($dir);
+          }
 
           // Restore Tar file
           chdir(dirname($dir)); // go to parent directory
